Postmerge fixes. WIP.

+++ layers/common_layers.py
@@ -132,7 +132,7 @@ class GravesAttention(nn.Module):
 
     def init_states(self, inputs):
         if self.J is None or inputs.shape[1] > self.J.shape[-1]:
-            self.J = torch.arange(0, inputs.shape[1]).to(inputs.device).expand([inputs.shape[0], self.K, inputs.shape[1]])
+            self.J = torch.arange(0, inputs.shape[1], dtype=torch.float32).to(inputs.device).expand([inputs.shape[0], self.K, inputs.shape[1]])
         self.attention_weights = torch.zeros(inputs.shape[0], inputs.shape[1]).to(inputs.device)
         self.mu_prev = torch.zeros(inputs.shape[0], self.K).to(inputs.device)
